Introduction
In this task we were required to implement a directed weighted graph with all the algorithms from the previous task as well as the graphical display (GUI) translated to Python. In addition we were required to compare execution times between our results in Java and Python.
This project purpose is to compare a Java code to its Python version. The Python version is this repository and the Java one is in link https://github.com/YehielSiri/DirectedWeightedGraph